backend/requirements.txt

fastapi>=0.115.0
uvicorn[standard]>=0.30.0

pydantic>=2.7.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1

motor>=3.5.0
pymongo>=4.8.0

pyjwt>=2.9.0
cryptography>=43.0.0

bcrypt>=4.1.0

requests>=2.32.0

pandas>=2.2.2
python-multipart>=0.0.9
openpyxl>=3.1.2